chore: move go code to go folder
