+++ .github/workflows/java.yaml
@@ -8,7 +8,7 @@ on:
       - WORKSPACE
   push:
     branches:
-      main
+      - main
 
 jobs:
   java:
@@ -20,7 +20,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ~/.cache/bazel
-          key: ${{ runner.os }}-bazel-cache
+          key: ${{ runner.os }}-bazel-cache-java
       - run: |
           bazel build //java:App
           bazel run //java:App
