+++ .github/workflows/python.yaml
@@ -8,7 +8,7 @@ on:
       - WORKSPACE
   push:
     branches:
-      main
+      - main
 
 jobs:
   python:
@@ -20,6 +20,6 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ~/.cache/bazel
-          key: ${{ runner.os }}-bazel-cache
+          key: ${{ runner.os }}-bazel-cache-python
       - run: |
           bazel run //python:main
