+++ go/cmd/BUILD.bazel
@@ -3,9 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "cmd_lib",
     srcs = ["main.go"],
-    importpath = "github.com/nakamasato/bazel-training/cmd",
+    importpath = "github.com/nakamasato/bazel-training/go/cmd",
     visibility = ["//visibility:private"],
-    deps = ["//uuid"],
+    deps = ["//go/uuid"],
 )
 
 go_binary(
