+++ go/uuid/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "uuid",
     srcs = ["uuid.go"],
-    importpath = "github.com/nakamasato/bazel-training/uuid",
+    importpath = "github.com/nakamasato/bazel-training/go/uuid",
     visibility = ["//visibility:public"],
     deps = ["@com_github_google_uuid//:uuid"],
 )
